re-alterando filtros e melhor organização de tudo
