+++ src/main/java/br/edu/unicesumar/folia/controller/boleto/BoletoRestController.java
@@ -1,13 +1,15 @@
 package br.edu.unicesumar.folia.controller.boleto;
 
+import br.edu.unicesumar.folia.controller.usuarioboleto.EmpresaBoletoFiltradoDTO;
+import br.edu.unicesumar.folia.controller.usuarioboleto.UsuarioBoletoFiltradoDTO;
+import br.edu.unicesumar.folia.controller.usuarioboleto.BoletoFiltroDTO;
+import br.edu.unicesumar.folia.controller.usuarioboleto.UsuarioBoletoListaDTO;
 import br.edu.unicesumar.folia.domain.boleto.Boleto;
 import br.edu.unicesumar.folia.domain.boleto.BoletoRepository;
 import br.edu.unicesumar.folia.domain.boleto.BoletoService;
-import br.edu.unicesumar.folia.domain.usuario.TipoUsuario;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
@@ -46,6 +48,7 @@ public class BoletoRestController {
         Page<BoletoListaDTO> boletos = boletoService.listarBoletosPorUsuario(uuid, pageable);
         return ResponseEntity.ok(boletos);
     }
+
     // Buscar boleto
     @GetMapping("/{uuid}")
     public ResponseEntity<Boleto> buscarBoleto(@PathVariable UUID uuid) {
@@ -57,15 +60,14 @@ public class BoletoRestController {
         }
     }
 
-
     // Atualizar status boleto
     @PutMapping("/{uuid}/status")
-    public ResponseEntity<Void> atualizarStatusBoleto(@PathVariable UUID uuid, @RequestParam String novoStatus) {
+    public ResponseEntity<String> atualizarStatusBoleto(@PathVariable UUID uuid, @RequestParam String novoStatus) {
         try {
             boletoService.atualizarStatusBoleto(uuid, novoStatus);
-            return new ResponseEntity<>(HttpStatus.OK);
+            return new ResponseEntity<>("Status do boleto atualizado com sucesso",HttpStatus.OK);
         } catch (EntityNotFoundException e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>("Boleto não encontrado",HttpStatus.NOT_FOUND);
         }
     }
 
@@ -73,6 +75,22 @@ public class BoletoRestController {
     public ResponseEntity<BoletoInformacoesDTO> informacoesBoletos(@PathVariable UUID uuid) {
         BoletoInformacoesDTO boletos = boletoService.listarBoletosPorEmpresa(uuid);
         return new ResponseEntity<>(boletos, HttpStatus.OK);
+    }
+
+    @GetMapping("/empresa/{empresaUuid}")
+    public ResponseEntity<List<UsuarioBoletoListaDTO>> obterBoletosPorEmpresa(@PathVariable UUID empresaUuid) {
+        List<UsuarioBoletoListaDTO> dtos = boletoService.obterDadosUsuariosPorEmpresa(empresaUuid);
+        return ResponseEntity.ok(dtos);
+    }
+
+    @PostMapping("/empresa/filtrar")
+    public ResponseEntity<List<EmpresaBoletoFiltradoDTO>> filtrarBoletosEmpresa(@RequestBody BoletoFiltroDTO filtro) {
+        return ResponseEntity.ok(boletoService.buscarComFiltroEmpresa(filtro));
+    }
+
+    @PostMapping("/usuario/filtrar")
+    public ResponseEntity<List<UsuarioBoletoFiltradoDTO>> filtrarBoletosUsuario(@RequestBody BoletoFiltroDTO filtro) {
+        return ResponseEntity.ok(boletoService.buscarComFiltroUsuario(filtro));
     }
 
 }
