+++ src/main/java/br/edu/unicesumar/folia/controller/usuarioboleto/BoletoFiltroDTO.java
@@ -1,23 +1,24 @@
 package br.edu.unicesumar.folia.controller.usuarioboleto;
 
-import br.edu.unicesumar.folia.domain.banco.Banco;
 import br.edu.unicesumar.folia.domain.boleto.Status;
-import br.edu.unicesumar.folia.domain.usuario.Usuario;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.time.LocalDate;
-import java.util.UUID;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class UsuarioBoletoFiltroDTO {
-    private String identificacao;
+public class BoletoFiltroDTO {
+    private String usuarioUUID;
+    private String empresaUUID;
+    private String banco;
     private String nome;
+    private String identificacao;
+    private String numero;
     private LocalDate dataInicialEmissao;
     private LocalDate dataFinalEmissao;
     private LocalDate dataInicialVencimento;
